waterfall: prevent tooltip from flashing in Firefox
For some reason the mouseover/mouseout event keeps triggering in Firefox
when hovering over a bar in the waterfall view, resulting in
continuously (dis)appearing tooltips and a bunch of requests to /steps.

Prevent the tooltip from generating mouse events to avoid this problem.

Reproducer: go to https://nine.buildbot.net/#/waterfall in Firefox
53.0.3 and hover over worker 446.